+++ src/modules/tools/saveSelection/store/gettersSaveSelection.js
@@ -4,32 +4,14 @@ import initialState from "./stateSaveSelection";
 const getters = {
     ...generateSimpleGetters(initialState),
     /**
-     * @param {Object} _ saveSelection store state.
-     * @param {Object} __ saveSelection store getters.
-     * @param {Object} ___ root state.
-     * @param {Object} rootGetters root getters.
-     * @returns {?[Number, Number]} The current center coordinates of the mapView.
-     */
-    centerCoordinates (_, __, ___, rootGetters) {
-        return rootGetters["Map/center"];
-    },
-    /**
      * @param {Object} state saveSelection store state.
-     * @param {Object} getters saveSelection store getters.
+     * @param {Object} _ saveSelection store getters.
+     * @param {Object} __ root state.
+     * @param {Object} rootGetters root getters.
      * @returns {String} The Url that can be copied by the user.
      */
-    url ({layerIds, layerVisibilities, layerTransparencies}, {centerCoordinates, zoomLevel}) {
-        return location.origin + location.pathname + "?layerIds=" + layerIds + "&visibility=" + layerVisibilities + "&transparency=" + layerTransparencies + "&center=" + centerCoordinates + "&zoomLevel=" + zoomLevel;
-    },
-    /**
-     * @param {Object} _ saveSelection store state.
-     * @param {Object} __ saveSelection store getters.
-     * @param {Object} ___ root state.
-     * @param {Object} rootGetters root getters.
-     * @returns {?[Number, Number]} The current zoomLevel of the mapView.
-     */
-    zoomLevel (_, __, ___, rootGetters) {
-        return rootGetters["Map/zoomLevel"];
+    url ({layerIds, layerVisibilities, layerTransparencies}, _, __, rootGetters) {
+        return location.origin + location.pathname + "?layerIds=" + layerIds + "&visibility=" + layerVisibilities + "&transparency=" + layerTransparencies + "&center=" + rootGetters["Map/center"] + "&zoomLevel=" + rootGetters["Map/zoomLevel"];
     }
 };